Removed unnecessary BasicErrorRenderer
refs: https://github.com/TryGhost/Ghost/commit/f685f2a8bfdf54803f00a2e6495f6edf836a1eab
refs: https://github.com/TryGhost/Ghost/commit/121ca53d823b119380c46807b2f889c6834c7a35

The BasicErrorRenderer was created as a fallback for when we needed to not render templates, which is
chiefly when we're trying to render a 404 for an image. Using a template puts us at risk of an infinite 404 loop
if the missing image is referenced in the 404 template.

As of 121ca53d, the HTMLErrorRenderer no longer uses templates - instead we serve a very simple HTML page.
This can be used instead of the BasicErrorRenderer, as it results in a properly formatted error.
Even when sending responses in plain text, the content type is returned as HTML and therefore having an
unformatted error makes no sense - if we really need a non-html format I guess there should be no body at all.