Removed archived Tiers from Portal display
refs https://github.com/TryGhost/Team/issues/1252

In order to ensure that archived Tiers are not selected to be shown in
Portal, both after they archived, and after they've been un-archived -
we need to persist this information to storage. This is currently done
via the `portal_products` setting.

The reasoning behind this is so that un-archiving a Tier doesn't modify
Portal settings without the user being aware.