Fixed paid signup notification message
refs https://github.com/TryGhost/Ghost/commit/9a76962ed0b85884c3df66a2ca27b8e9a8cac970

As part of the work of automatically logging members in after payment, we added a new action type for signup - `signup-paid` - which allows custom message for a member that is already signed up and doesn't need to confirm their subscription. The new action - `signup-paid` - wasn't added to Portal notification handler which then fallbacks to showing just `Success` as the message. This change -

- handles new `signup-paid` action type allowing Portal to show proper success message