Readded notEnoughPermission message in post model (#14759)
no issue

The message was removed somewhere, but the git history looks to be modified so I couldn't find the reference to when and why it was removed (the message seems to be added and removed in the same commit).

Added in: https://github.com/tryghost/Ghost/commit/04f19d362d25afa122e024034dcd97b0c068e060
No reference to where it was removed. But after the same commit above, it wasn't present any longer.